chore: move normalizeCourseSetting to utils/course.ts
